Add test cases

+++ Team28/Code28/src/spa/src/QPS/Evaluator/Tuple.h
@@ -67,6 +67,6 @@ public:
 	}
 
 	Value getValue(int index) {
-		return this->values[index];
+		return this->values[(size_t) index];
 	}
 };
